requirements.txt
africastalking==2.0
asgiref==3.9.1
black==25.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
coverage==7.10.7
Django==5.2.6
djangorestframework==3.16.1
dotenv==0.9.9
flake8==7.3.0
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
playwright==1.55.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
pytest-base-url==2.1.0
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-playwright==0.7.1
python-decouple==3.8
python-dotenv==1.1.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.5
responses==0.25.8
schema==0.7.7
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.5.0
